backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.6
sqlalchemy[asyncio]>=2.0.35
asyncpg>=0.29.0
alembic>=1.13.3
openai>=1.50.0
python-jose[cryptography]==3.3.0
bcrypt>=4.0.0
pydantic-settings>=2.5.2
pydantic>=2.9.2
httpx>=0.27.2
pytest>=8.3.3
pytest-asyncio>=0.24.0
python-multipart>=0.0.12
python-dateutil>=2.9.0
email-validator>=2.2.0
